'Lite' Optimizations and Tweaks for Win32CYG Build

+++ src/edacs-bch3.c
@@ -480,7 +480,8 @@ decode_bch()
 	}
 }
 
-long int messagepp = 0x0;
+//bugfix on 32-bit cygwin, probably overflowed here
+unsigned long long int messagepp = 0x0;
 
 //very simplified version, just to encode, get frame and compare
 unsigned long long int edacs_bch (unsigned long long int message)
